+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+v1.4
+----
+- Added script to import SQL file to server (07/08/2019)
+- Added list of known bots (07/08/2019)
+- Added list of known file extensions (07/08/2019)
+
 v1.3
 ----
 - Made use of `git reset --hard origin/master` to force overwrite of local changes (03/08/2019)